# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lara Clinical Med",
        "version": "0.1"
    },
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Listar pacientes",
                "description": "Endpoint para listar todos os pacientes",
                "operationId": "0df0d65f049161931654c3e188ca8f93",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sucesso"
                                        },
                                        "patients": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "João da Silva"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "11987654321"
                                                    },
                                                    "identity": {
                                                        "type": "string",
                                                        "example": "11122233344"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nenhum paciente foi encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Criar um novo paciente",
                "description": "Endpoint para criar um novo paciente",
                "operationId": "2149729f1bf1af8877352f04be3a683f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do paciente",
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "phone": {
                                        "description": "Telefone do paciente",
                                        "type": "string",
                                        "example": "11987654321"
                                    },
                                    "identity": {
                                        "description": "CPF do paciente",
                                        "type": "string",
                                        "format": "string",
                                        "example": "11122233344"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente criado com sucesso."
                                        },
                                        "patient": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João da Silva"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "11987654321"
                                                },
                                                "identity": {
                                                    "type": "string",
                                                    "example": "11122233344"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível criar o paciente."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/patients/{patientId}": {
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Atualizar um paciente",
                "description": "Endpoint para atualizar um paciente existente",
                "operationId": "1762fe131b200621915a2daa7cee79d2",
                "parameters": [
                    {
                        "name": "patientId",
                        "in": "path",
                        "description": "ID do paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do paciente",
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "phone": {
                                        "description": "Telefone do paciente",
                                        "type": "string",
                                        "example": "11987654321"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paciente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente atualizado com sucesso."
                                        },
                                        "patient": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "João da Silva"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "(11) 98765-4321"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Não foi possível atualizar o paciente."
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Validation error message"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}